Files modified: 2
Modified 2 files:number/NumberModifier.swiftnumber/int/IntModifier.swift

+++ number/int/IntModifier.swift
@@ -1,8 +1,9 @@
 class IntModifier {
     /*
-     * 
+     * Directly replaces a with b, and b with a
+     * NOTE: Modifies the original a and b Doubles
      */
-    class func swap(inout a:Double, inout _ b:Double) {
+    class func swap(inout a:Int, inout _ b:Int) {
         ObjectModifier.swapObjects(&a, &b)
     }
 }
